Fixes #218 Titles show encoding problems
This commit fixes #218, where a title of an article was html encoded.
Now the title is html decoded before storing it as article with the help
of the method Html.fromHtml()

E.g. before:
Using Let&#039;s Encrypt with Kerio Connect
Now is saved:
Using Let's Encrypt with Kerio Connect