Fixes #221 Deleting an articles throws a 404 error
This commit fixes #221 and some more issue w.r.t. toggle actions on an
article. In wallabag v2 if we request a toggle action, e.g. /star/123,
and if we are not logged in anymore, we get redirected to that action
after successful login. But our app then did re-toggle that action.
Therefore the action had been called twice. When deleting an article
this resulted in a 404. When starring or archiving, this toggled the
action twice --> resulting in no action.

So now if we recognize that we executing an action results in a login
page, we request a new login page, do a clean login and afterwards
recall the action request. This way we avoid getting redirected to the
action request.