improve UI in preference screen by using categories
This commit adds preference categories as indication in which of the
setting's sub menu you currently are. This was already done for the
"Connection" menu, but not for the other tree menu points.
This commit moves the "Delete cookies" option visually (not
syntactically) from the HTTP Basic Auth category to the "Advanced
connection settings". That way the cookies option is not linked to the
HTTP basic auth options anymore.