Use right color for dark contrast theme
Use a real black instead of grey in the dark contrast theme for the
app bar. Also remove the black square in the navigation bar for the selected item
and instead use grey to highlight it.
This was done by changing the primary color of the theme and not
changing the background attribute in the theme to not always be black.